# server/src/routes/camera.js
const express = require('express');
const router = express.Router();
const logger = require('../utils/logger');
const { validateRequest } = require('../utils/validation');
const Joi = require('joi');

// Validation schemas
const cameraSchema = Joi.object({
  name: Joi.string().required(),
  location: Joi.object({
    latitude: Joi.number().min(-90).max(90).required(),
    longitude: Joi.number().min(-180).max(180).required(),
    address: Joi.string().optional()
  }).required(),
  status: Joi.string().valid('active', 'inactive', 'maintenance').default('active'),
  configuration: Joi.object({
    detection_enabled: Joi.boolean().default(true),
    image_quality: Joi.string().valid('low', 'medium', 'high').default('medium'),
    upload_interval: Joi.number().min(1).max(3600).default(60)
  }).optional()
});

/**
 * POST /api/v1/camera
 * Register a new camera
 */
router.post('/', validateRequest(cameraSchema), async (req, res) => {
  try {
    const cameraData = req.validatedData;
    
    logger.info('Registering new camera', {
      name: cameraData.name,
      location: cameraData.location
    });

    // TODO: Save to database
    const mockCamera = {
      id: 'cam-' + Date.now(),
      ...cameraData,
      created_at: new Date().toISOString(),
      updated_at: new Date().toISOString()
    };

    res.status(201).json({
      success: true,
      message: 'Camera registered successfully',
      data: mockCamera
    });
  } catch (error) {
    logger.error('Error registering camera:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

/**
 * GET /api/v1/camera
 * Get all cameras
 */
router.get('/', async (req, res) => {
  try {
    const { status, limit = 50, offset = 0 } = req.query;

    // TODO: Implement database query with filters
    const mockCameras = [
      {
        id: 'cam-001',
        name: 'Main Entrance Camera',
        location: {
          latitude: 40.7128,
          longitude: -74.0060,
          address: '123 Main St, New York, NY'
        },
        status: 'active',
        configuration: {
          detection_enabled: true,
          image_quality: 'high',
          upload_interval: 30
        },
        created_at: new Date().toISOString(),
        updated_at: new Date().toISOString()
      }
    ];

    res.json({
      success: true,
      data: mockCameras,
      pagination: {
        limit: parseInt(limit),
        offset: parseInt(offset),
        total: mockCameras.length
      }
    });
  } catch (error) {
    logger.error('Error fetching cameras:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

/**
 * GET /api/v1/camera/:id
 * Get specific camera by ID
 */
router.get('/:id', async (req, res) => {
  try {
    const { id } = req.params;

    // TODO: Implement database query
    const mockCamera = {
      id,
      name: 'Main Entrance Camera',
      location: {
        latitude: 40.7128,
        longitude: -74.0060,
        address: '123 Main St, New York, NY'
      },
      status: 'active',
      configuration: {
        detection_enabled: true,
        image_quality: 'high',
        upload_interval: 30
      },
      created_at: new Date().toISOString(),
      updated_at: new Date().toISOString()
    };

    res.json({
      success: true,
      data: mockCamera
    });
  } catch (error) {
    logger.error('Error fetching camera by ID:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

/**
 * PUT /api/v1/camera/:id
 * Update camera configuration
 */
router.put('/:id', validateRequest(cameraSchema), async (req, res) => {
  try {
    const { id } = req.params;
    const updateData = req.validatedData;
    
    logger.info('Updating camera configuration', { id, updateData });

    // TODO: Update in database
    const mockUpdatedCamera = {
      id,
      ...updateData,
      updated_at: new Date().toISOString()
    };

    res.json({
      success: true,
      message: 'Camera updated successfully',
      data: mockUpdatedCamera
    });
  } catch (error) {
    logger.error('Error updating camera:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

/**
 * DELETE /api/v1/camera/:id
 * Deactivate camera
 */
router.delete('/:id', async (req, res) => {
  try {
    const { id } = req.params;
    
    logger.info('Deactivating camera', { id });

    // TODO: Update status in database

    res.json({
      success: true,
      message: 'Camera deactivated successfully'
    });
  } catch (error) {
    logger.error('Error deactivating camera:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

module.exports = router;