# server/src/routes/detection.js
const express = require('express');
const router = express.Router();
const logger = require('../utils/logger');
const { validateRequest } = require('../utils/validation');
const Joi = require('joi');

// Validation schemas
const detectionSchema = Joi.object({
  camera_id: Joi.string().required(),
  timestamp: Joi.date().iso().required(),
  license_plate: Joi.string().pattern(/^[A-Z0-9]{1,10}$/).required(),
  confidence: Joi.number().min(0).max(1).required(),
  image_url: Joi.string().uri().optional(),
  location: Joi.object({
    latitude: Joi.number().min(-90).max(90).optional(),
    longitude: Joi.number().min(-180).max(180).optional()
  }).optional(),
  vehicle_info: Joi.object({
    make: Joi.string().optional(),
    model: Joi.string().optional(),
    color: Joi.string().optional(),
    type: Joi.string().optional()
  }).optional()
});

/**
 * POST /api/v1/detection
 * Receive detection data from edge cameras
 */
router.post('/', validateRequest(detectionSchema), async (req, res) => {
  try {
    const detectionData = req.validatedData;
    
    logger.info('Received detection data', {
      camera_id: detectionData.camera_id,
      license_plate: detectionData.license_plate,
      confidence: detectionData.confidence
    });

    // TODO: Save to database
    // TODO: Process analytics
    // TODO: Send notifications if needed

    res.status(201).json({
      success: true,
      message: 'Detection data received successfully',
      id: 'temp-id-' + Date.now() // TODO: Use actual database ID
    });
  } catch (error) {
    logger.error('Error processing detection data:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

/**
 * GET /api/v1/detection
 * Get detection history with optional filters
 */
router.get('/', async (req, res) => {
  try {
    const { 
      camera_id, 
      license_plate, 
      start_date, 
      end_date, 
      limit = 50, 
      offset = 0 
    } = req.query;

    // TODO: Implement database query with filters
    const mockData = [
      {
        id: '1',
        camera_id: 'cam-001',
        timestamp: new Date().toISOString(),
        license_plate: 'ABC123',
        confidence: 0.95,
        image_url: 'https://example.com/image1.jpg'
      }
    ];

    res.json({
      success: true,
      data: mockData,
      pagination: {
        limit: parseInt(limit),
        offset: parseInt(offset),
        total: mockData.length
      }
    });
  } catch (error) {
    logger.error('Error fetching detection data:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

/**
 * GET /api/v1/detection/:id
 * Get specific detection by ID
 */
router.get('/:id', async (req, res) => {
  try {
    const { id } = req.params;

    // TODO: Implement database query
    const mockData = {
      id,
      camera_id: 'cam-001',
      timestamp: new Date().toISOString(),
      license_plate: 'ABC123',
      confidence: 0.95,
      image_url: 'https://example.com/image1.jpg'
    };

    res.json({
      success: true,
      data: mockData
    });
  } catch (error) {
    logger.error('Error fetching detection by ID:', error);
    res.status(500).json({
      success: false,
      error: 'Internal server error'
    });
  }
});

module.exports = router;